requirements.txt
langchain-community==0.3.1 
langchain_text_splitters==0.3.5 
langchain-groq==0.2.3 
langchain==0.3.17 
youtube-transcript-api==0.6.3 
pytube==15.0.0
sentence_transformers==3.3.1
python-dotenv